# Implement optimized index calculations for collapse(4) and higher